✅ [Logs] move public API tests in logs.entry
